% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lheatmap.r
\name{Xheatmap}
\alias{Xheatmap}
\title{for your own heatmap at anywhere}
\usage{
Xheatmap(table, coordinate, rowname = TRUE, picnum)
}
\arguments{
\item{table, data}{matrix}

\item{coordinate, the}{image's site to plot}

\item{rowname, an}{boolean to mean if table has the row names}
}
\description{
for your own heatmap at anywhere
}
